src/test/java/nextstep/subway/unit/line/LineServiceTest.java
package nextstep.subway.unit.line;

import nextstep.subway.dto.LineRequest;
import nextstep.subway.dto.LineResponse;
import nextstep.subway.dto.SectionRequest;
import nextstep.subway.service.LineService;
import nextstep.subway.dto.StationRequest;
import nextstep.subway.dto.StationResponse;
import nextstep.subway.service.StationService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import java.util.stream.Collectors;

import static org.assertj.core.api.Assertions.assertThat;

@DisplayName("구간 서비스 단위 테스트 without mock")
@SpringBootTest
@Transactional
public class LineServiceTest {

    @Autowired
    StationService stationService;

    @Autowired
    LineService lineService;

    StationResponse 강남역_생성_응답;
    StationResponse 역삼역_생성_응답;
    StationResponse 선릉역_생성_응답;
    LineResponse 이호선_생성_응답;

    @BeforeEach
    void setUp() {
        강남역_생성_응답 = stationService.saveStation(new StationRequest("강남역"));
        역삼역_생성_응답 = stationService.saveStation(new StationRequest("역삼역"));
        선릉역_생성_응답 = stationService.saveStation(new StationRequest("선릉역"));
        이호선_생성_응답 = lineService.save(new LineRequest("이호선", "green", 강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 10));
    }

    @DisplayName("구간을 추가한다.")
    @Test
    void addSection() {
        // when
        lineService.addSection(이호선_생성_응답.getId(), new SectionRequest(역삼역_생성_응답.getId(), 선릉역_생성_응답.getId(), 10));

        // then
        assertThat(lineService.findLineById(이호선_생성_응답.getId())).isNotNull();
        assertThat(lineService.findLineById(이호선_생성_응답.getId()).getId()).isNotNull();
        assertThat(lineService.findLineById(이호선_생성_응답.getId()).getStations().stream()
                .map(s -> s.getId())
                .collect(Collectors.toList())
        ).containsExactly(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 선릉역_생성_응답.getId());
    }

    @DisplayName("구간을 삭제한다.")
    @Test
    void deleteSection() {
        // given
        lineService.addSection(이호선_생성_응답.getId(), new SectionRequest(역삼역_생성_응답.getId(), 선릉역_생성_응답.getId(), 10));

        // when
        lineService.deleteSection(이호선_생성_응답.getId(), 선릉역_생성_응답.getId());

        // then
        assertThat(lineService.findLineById(이호선_생성_응답.getId())).isNotNull();
        assertThat(lineService.findLineById(이호선_생성_응답.getId()).getId()).isNotNull();
        assertThat(lineService.findLineById(이호선_생성_응답.getId()).getStations().stream()
                .map(s -> s.getId())
                .collect(Collectors.toList())
        ).containsExactly(강남역_생성_응답.getId(), 역삼역_생성_응답.getId());
    }
}
